requirements.txt
# Core dependencies
aiohttp>=3.8.0
asyncio>=3.4.3
python-dotenv>=0.19.0
pydantic>=1.9.0
loguru>=0.5.3

# Solana dependencies
solana>=0.30.0
anchorpy>=0.18.0

# Monitoring and logging
prometheus-client>=0.12.0

# Testing
pytest>=6.2.5
pytest-asyncio>=0.16.0
pytest-cov>=2.12.0

# Development
black>=21.12b0
isort>=5.10.0
flake8>=4.0.0
mypy>=0.910

# AI and ML
torch>=1.10.0
transformers>=4.15.0
numpy>=1.21.0
pandas>=1.3.0

# Blockchain
web3>=5.24.0

# Testing
pytest>=6.2.5
pytest-asyncio>=0.16.0
pytest-cov>=2.12.0

# Development
black>=21.12b0
isort>=5.10.0
flake8>=4.0.0
mypy>=0.910 